Use of Guid as a session identifer #427
