+++ Tests/Titanium.Web.Proxy.UnitTests/WinAuthTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Titanium.Web.Proxy.Http;
 using Titanium.Web.Proxy.Network.WinAuth;
 
 namespace Titanium.Web.Proxy.UnitTests
@@ -10,7 +11,7 @@ namespace Titanium.Web.Proxy.UnitTests
         [TestMethod]
         public void Test_Acquire_Client_Token()
         {
-            string token = WinAuthHandler.GetInitialAuthToken("mylocalserver.com", "NTLM", Guid.NewGuid());
+            string token = WinAuthHandler.GetInitialAuthToken("mylocalserver.com", "NTLM", new InternalDataStore());
             Assert.IsTrue(token.Length > 1);
         }
     }
